requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.7
google-api-python-client==2.108.0
google-generativeai==0.3.2
notion-client==2.2.1
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
perplexity==0.0.1 